% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.housing.access-package.R
\docType{package}
\name{dc.acs.housing.access-package}
\alias{dc.acs.housing.access}
\alias{dc.acs.housing.access-package}
\title{dc.acs.housing.access: Get Housing Access Data From SDAD Data Commons}
\description{
Allows user to easily get housing access data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia

}
\keyword{internal}
